% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Make a forest plot}
\usage{
forest_plot(tbl, group_by_type = TRUE)
}
\arguments{
\item{tbl}{A tibble.}

\item{group_by_type}{Whether to group by effect size type.}
}
\description{
Makes a forest plot.
}
