% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quincunx_pgs_report.R
\name{quincunx_pgs_report}
\alias{quincunx_pgs_report}
\title{Make a PGS Report}
\usage{
quincunx_pgs_report(
  scores_obj,
  file = basename(tempfile(pattern = "pgs_report_", fileext = ".html"))
)
}
\arguments{
\item{scores_obj}{A PGS scores object.}

\item{file}{A file path to be used for saving the report.}
}
\description{
Makes a PGS report.
}
